migrate to v2 embedding
